requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
safetensors>=0.3.0
opencv-python>=4.8.0
Pillow>=9.5.0
numpy>=1.24.0
scipy>=1.10.0
mediapipe>=0.10.0
flask>=2.3.0
flask-cors>=4.0.0
werkzeug>=2.3.0
requests>=2.31.0
tqdm>=4.65.0
facexlib>=0.3.0
basicsr>=1.4.2
gfpgan>=1.3.8
codeformer>=0.1.0
insightface>=0.7.3
onnxruntime>=1.15.0
xformers>=0.0.20
peft>=0.4.0
accelerate>=0.20.0
datasets>=2.10.0
psutil>=5.9.0 